# Fix `no-invalid-double-slash-comments` reported ranges