--- a/.changeset/giant-days-play.md
+++ b/.changeset/giant-days-play.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `no-invalid-double-slash-comments` reported ranges