--- a/lib/rules/no-invalid-double-slash-comments/__tests__/index.mjs
+++ b/lib/rules/no-invalid-double-slash-comments/__tests__/index.mjs
@@ -1,6 +1,8 @@
 import rule from '../index.mjs';
 const { messages, ruleName } = rule;
 
+import { stripIndent } from 'common-tags';
+
 testRule({
 	ruleName,
 	config: [true],
@@ -17,6 +19,18 @@ testRule({
 		{
 			code: 'a { background: url(//foo.com/bar.png) }',
 			description: 'url with double slash',
+		},
+		{
+			code: stripIndent`
+				/* stylelint-disable-next-line no-invalid-double-slash-comments */
+				.a, // Comment 1
+				/* stylelint-disable-next-line no-invalid-double-slash-comments */
+				.b, // Comment 2
+				.c {
+					color: red;
+				}
+			`,
+			description: 'configuration commands interleaved with double slash comments',
 		},
 	],
 
@@ -74,6 +88,33 @@ testRule({
 			column: 1,
 			endLine: 1,
 			endColumn: 13,
+		},
+		{
+			code: stripIndent`
+				.a, // Comment 1
+				/* foo */
+				.b, // Comment 2
+				.c {
+					color: red;
+				}
+			`,
+			description: 'regular comments interleaved with double slash comments',
+			warnings: [
+				{
+					message: messages.rejected,
+					line: 1,
+					column: 5,
+					endLine: 2,
+					endColumn: 1,
+				},
+				{
+					message: messages.rejected,
+					line: 3,
+					column: 5,
+					endLine: 4,
+					endColumn: 1,
+				},
+			],
 		},
 	],
 });