--- a/lib/rules/no-invalid-double-slash-comments/index.cjs
+++ b/lib/rules/no-invalid-double-slash-comments/index.cjs
@@ -2,6 +2,7 @@
 // please instead edit the ESM counterpart and rebuild with Rollup (npm run build).
 'use strict';
 
+const getRuleSelector = require('../../utils/getRuleSelector.cjs');
 const report = require('../../utils/report.cjs');
 const ruleMessages = require('../../utils/ruleMessages.cjs');
 const validateOptions = require('../../utils/validateOptions.cjs');
@@ -37,7 +38,7 @@ const rule = (primary) => {
 		});
 
 		root.walkRules((ruleNode) => {
-			const selectors = (ruleNode.raws?.selector?.raw ?? ruleNode.selector).split(',');
+			const selectors = getRuleSelector(ruleNode).split(',');
 
 			let ruleStringified;
 			let index = 0;