--- a/lib/rules/no-invalid-double-slash-comments/index.mjs
+++ b/lib/rules/no-invalid-double-slash-comments/index.mjs
@@ -33,9 +33,9 @@ const rule = (primary) => {
 		});
 
 		root.walkRules((ruleNode) => {
-			// ruleNode.selectors elements are trimmed
-			const selectors = ruleNode.selector.split(',');
-			const ruleStringified = ruleNode.toString();
+			const selectors = (ruleNode.raws?.selector?.raw ?? ruleNode.selector).split(',');
+
+			let ruleStringified;
 			let index = 0;
 
 			for (const value of selectors) {
@@ -44,6 +44,8 @@ const rule = (primary) => {
 				if (selector.startsWith('//')) {
 					const offset = value.length - selector.length;
 					const i = index + offset;
+
+					ruleStringified ??= ruleNode.toString();
 					const lines = ruleStringified.slice(i).split('\n');
 					let comment = lines[0] ?? '';