--- a/lib/rules/no-invalid-double-slash-comments/index.mjs
+++ b/lib/rules/no-invalid-double-slash-comments/index.mjs
@@ -1,3 +1,4 @@
+import getRuleSelector from '../../utils/getRuleSelector.mjs';
 import report from '../../utils/report.mjs';
 import ruleMessages from '../../utils/ruleMessages.mjs';
 import validateOptions from '../../utils/validateOptions.mjs';
@@ -33,7 +34,7 @@ const rule = (primary) => {
 		});
 
 		root.walkRules((ruleNode) => {
-			const selectors = (ruleNode.raws?.selector?.raw ?? ruleNode.selector).split(',');
+			const selectors = getRuleSelector(ruleNode).split(',');
 
 			let ruleStringified;
 			let index = 0;