<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

See https://github.com/stylelint/stylelint/issues/7904

> Is there anything in the PR that needs further explanation?

I also noticed that rules were eagerly serialized, this is quite expensive.

This change:

```
;node benchmark-rule.mjs no-invalid-double-slash-comments 'true'                               
Warnings: 0
Mean: 95.35613458823532 ms
Deviation: 9.862092026552252 ms
```

Main:

```
;node benchmark-rule.mjs no-invalid-double-slash-comments 'true'
Warnings: 0
Mean: 118.80464825 ms
Deviation: 23.471924727353816 ms
```